// apps/web-panel-vue/src/views/site/types.d.ts
export interface SslInfoProps {
  isCertInfo: boolean
  https: boolean
  type: number
  key: string
  csr: string
  index: string
  oid: number
  dns: string[]
  subject: string
  issuer: string
  endtime: string
  push: AnyObject
  pushAlarm: string | boolean
  isSupportRenewal: boolean
  isStart: boolean
  notAfter: string
  authType: string
}
// ...
export interface LetsTableProps { 
  auth_to: string
  auth_type: string
  cert_timeout: number
  certificate_url: string
  domains: string[]
  endDay: number
  expires: number
  index: string
  next_retry_time: number
  renew_time: number
  retry_count: number
  save_path: string
  status: string
}